readSBML SBMLImporter now takes a CFunctionDB object as second argument.
The power function in ASTNode trees is now converted into the power operator
so that copasi can parse the resulting function string.

The divide and multiply operation by the compartment volume upon import
and export is only done when the volume is not 1. If there already is the
opposite reaction from an earlier export, this is dropped rather than adding
another level.