First draft of the userguide. So far only the chapter about the GUI is written.
I also checked in the html and the tex file generated from the source files since it is rather complicated to generate them and it will not
work on most machines. I will have to write some kind of readme that explains the steps necessary to create all files.

A makefile that may work on some machines will be written as well.