Fixed the segfaults by not calling copy constructors of nodes when converting from an expression to a function tree.
Changed data for one function from acosh to arccosh. The type names and real
names and SBML names etc. is rather confusing since in some function the
textual representation is different from the type which might be again different
from what it is in SBML.

Improved the wrapper script for the test suite a bit.