Some progress on roundtripping SBML files.
Most of the objects should survive roundtrips already. Some problems might
be encountered when models with events or rules are read, modified and written
again.

The conversion of the KineticLaw needs to be rewritten to use the new function
tree. When I do that, I will try to get rountripping for those as well.