The exporter seems to work now. I could run the SBML semantic testsuite on files that were imported into copasi, written out as CPS file, read in again and exported as SBML file. All tests but two passed. Those two tests were sensitive to the order of the metabolites in the file (see Bug 333).
In order to be able to associate certain copasi output with entities in
an SBML file, we introduced a new attribute sbmlid to compartments,
metabolites, model values, reactions and functions. This attribute stores
the id of the corrsponding SBML entities and is used as the id for those
SBML entities on export.

This helps with running the SBML testsuite on exported files. But this feature
will soon be needed in other tools as well (e.g. Sycamore, or Katja Wegners
und Ursula Rosts vizualisation tools).