Another wrapper for the semantic test suite.
This one I use to test the SBML exporter.

It reads in SBML file, writes it out as a Copasi file, reads the Copasi
file back in and exports it again as an SBML file.

This SBML file is now handled as in the original wrapper. It is imported and
written out as a Copasi file again and to this copasi file, the task
description is added as well as a report definition. This file is then used
to create the numeric results for the test suite.