Updated most of the document now, only the Multiple Tasks are missing.
Added new sections for the Lyapunov Task and the C and Madonna Export.

Added some paragraphs about the hybrid solver but I was not sure if RK 5th
order was correct or if it was 3rd order.

I guess it would be good if Jrgen could write a little bit more about the
hybrid and write the methods section for it.

Natalia could provide something for the C and Madonna export.