Almost finished with adding all error explanations from Stefan. There are still one or two where I am not sure what they actually mean or how I can force them.
I also could not force any error when trying to read invalid configuration
files or configurations files from invalid places etc.

I have to create an SBML model that is explicitely time dependant to see if
we import those correctly. I am not sure if will will have time for that
before the release.

Some general explanations are still missing for eac error section, especially
on how the functin parser and lexer work so that the error messages are
understandable.

The whole file probably needs to be refromatted because I don't think a table
will look good for this.