Added some code to handle import of time dependent SBML models.
Accoring to my debbugging session, everything should be fine,
I debugged everything until after the test model for bug 568.

The reaction tree for function_4_reaction looks as I would expect it and
the only reaction in the model points to this reaction.

When I open the function widget and go to the function_4_reaction function,
the mathML display is wrong because it shows a recursive function call.
When I click the edit button, the correct infix shows up and when I click commit
again, the MathML display is wrong again.

When I select the only reaction in the tree, Copasi crashes.