Added some methods to check if a model can be exported to a given version of SBML.
Checks for SBML Level2 Version1 are in place, others will be added.

The warning that comes up if the model can not be exported completely has been
placed after the file selection on purpose since I expect future file selection
dialogs to ask the user about the version of SBML he wants to export to.
Since the checks depend on this, the warning shows up after the user has
selected the file.