Fixed a bug in the import of rules for compartments and species.
Updated the testlist for the semantic testsuite to include the rule tests.

It looks as if we pass all basic rules tests, although the test script
reports two failures because it has an incorrect value of 1.0e-14 instead of 0
for some species, but I will have to double check if this is the case.

I also added one more test to the functions section which contains a rule
and that one fails as well and I don't know why yet, maybe one of the input
files is not correct.