Some changes to the SBMLExporter API. Some functions now take a CCopasiDataModel object as an argument instead of a CModel. Other function take an additional CCopasiDataModel object. This made a small change in CCopasiDataModel necessary.
This allowed me to get rid of all references to CCopasiDataModel::Global
as Sven has suggested.

Changed a vector into a set to make removing of objects more efficient.

Fixes:

- Clean up the copasi2sbmlmap upon export.
- Remove rules that are no longer there
- Export rules for compartments and species (I actually thought that this
was already the case, but they only got exported due to a bug where in the exporter
that did not remove obsolete rules).