Fixed a bug in the simplify method for CNormalLogicalItem.
Rewrote parts of the simplify of CNormalLogical. Now it has one bug less and
is shorter and simpler.

Unfortunatelly I think the comparison of logical expressions only works if
they are converted to their canonical form which means generating the full
table. Right now the expressions that results from simplification doesn't seem
to be unique.