Fixed a number of bugs in the importer and exporter. e.g. FunctionDefinitions with explicit time dependency were imported incorrectly so far. , time was imported as a the moel object itself instead of as a reference to the time object of the model. This lead to incorrect cps files being written from those models. The method for finding used functions and ordering them was incorrect because it falsly assumed a loop if a function was used twice within the same tree.
With these fixes I could import all but 4 models from biomodels, the remaining four had stoichiometric
expressions or delay functions which are not supported yet.
I wrote CPS files for all of the remaining 146 models and reexported those as SBML to test the exporter.
While reexporting I got an error message from the XML file reader about some miriam element for some,
but the rest could be reexported without crashes or error messages.

This does not mean that the imported or reexported models are correct however.