Rewrite of the simplification of CEvaluationNode based trees. Hopefully this code will not run into the same problems as the old one.
Now it is all in one place and the methods in the individual CEvaluationNode
classes can evantually be deleted.

Right now the new code is untested, I only know that it compiles. But since
tests are already in place, this should be fast.

Also, the new code is highly inefficient since the tree that is simplified
is copied over and over again, but this should be easily fixed once the code
works.