Updated the schema to include events I don't know if it is complete. MIRIAM annotations for events and event elements might be missing.
The schame validates against the Schema DTD, so I am pretty sure that at least the syntax is correct.
In order to be able to validate COPASI files against the COPASI schema, I had to
change the pattern expression for cNAME from "[^\t\n\r]+" to "[\S ]+".
Out of some reason \t, \r and \n were interpreted as t, r and n respectively by xmllint
and all names that included those characters did not pass validation.

I am not sure if this is a bug in xmllint or in the old expression. But the new expression should be equivalent
to the old one.