Enlarged the default node size.
Create a dummy time series if no simulation has been run.
Unfortunatelly the widget does not recognize when a simulation is run afterwards.
I will have to add some way to let the user reload new time series data.

The scaling in CQLayoutMainWindow is now always done from 0.0 to 1.0 which will make it easier to
add custom gradients for the color code mode later on.