Change the way objects are looked up from common names.
I removed the ObjectFromName functions from CCopasiContainer.
Now there is one method called ObjectFromName in CCopasiDataModel which only takes
a common name and tries to find the object. This method is not static.

Another method called ObjectFromName is in CCopasiObject. It takes a common name
and a list of containers.
If the object is not found ion the given containers, it is searched in the parent datamodel and if
it is not found there, it is searched in the function database of the root container.

The root container itself is not searched right now.