I changed the default name for CCopasiDataModel back to Root.
If it is set to something else, the getObject Method will no longer work for
old common names, as they are currently found in COPASI files.

This is not a big problem, but if we want to change the name of the datamodel, we have
to at least change the getObject Method in CCopasiContainer to deal with this.