Programming overview for the copasi backend. This document is based on the tex file that has so far been distributed with the bindings. I converted it to the DocBook format and tried to write it in a way that the text is language agnostic and that only the programming examples are language specific. Each element that is language specific has an attribute called condition.
I had planned to include this file as many times as there are languages and set
a switch before each import to tell the XSLT processor which version to generate.
Unfortunatelly I could not find a way to control the XSLT processor from the document,
obviously parameters can only be set in the stylesheet.

The only way that seems to work is to make this a two step process.
In the first step, the XSLT processor generates x files for the different languages and
those files have to be imported. The drawback here is that when a new language is added,
the build system as well as the document has to be adjusted and we need a second XSLT stylesheet
that can generate the different input files from the copasi_bindings.xml file..