Small bugfix for the unit handling.
Also changed the way invalid units are handled. So far if the unit did not fit,
we refused to load the model. Now all that happens is that the user gets a warning.

Since units are only for display, the worst that can happen is that COPASI will
not display the same units as are defined in the SBML file.