Added some new warnings about missing or unhandled units in SBML L3 documents.
Some changes to the way parameters in kinetic laws are handled by libsbml (although I know now that this chnage would not have been necessary).

Started implementing handling of references to species references.

Small change to the way the useValuesFromTriggerTime is imported.

Changed the required libsbml version in the ifdef statements. The SBML 3 support code
now needs libsbml 4.1 or higher.
The old version only said that it needs libsml greater 3.4.1, but since most things won't work with libsbml 4.0.1
this did not make sense.