New code to handle new event features in SBML Level 3. There now will be warning messages if events with priorities are imported. On export all priorities will be deleted from events.
There  are new warnings about the persistent flag on event triggers and the
initialValue flag although on event triggers.
If those flags have a value of false, a warning is issued. On export, the flags are
always set to true.

If COPASI tries to import an SBML L3 file with a reuqired packages, we refuse to load.