Added a new method for creating screenshots to the animation GUI. This method can handle arbitrary parts of the layout and it can create screenshots for all animation frames in one go.
The method uses a multisample framebuffer to do the rendering and the rendering is
done in small tiles. Since the method now uses Qts mechanisms to set up the framebuffer,
it should hopefully work with more hardware and drivers. Even those without support
for multisampling.

The new method is currently not enabled because it only makes sense with
a new dialog for making the settings which I still have to write.

The nice thing is that I could use the method to create png files for the complete animation
and convert this to a movie file in an external program.