Ran into another problem that I don't know how to fix.
Somehow all function calls that take a reference to a size_t objects are not mapped
to the corresponding integer type in R, but R expects an object that extends
point to size_t in the wrapped calls. But first of all, this is inconvenient
and second, R does not know how to create an object of type size_t.

So right now my example fails where I try to get a parameter from a parameter group using the index
of the parameter.