Some small bugfixes to the wrapper files for CTimeSeries and CEvaluationTree.
Tried to fix the function add bug for the wrapped CFunctionDB class, but at least for Java
the fix as it is documented in the SWIG documentation does not seem to work. I have no clue why
since the fix keeps the object referenced and should avoid it being collected by the garbage collector.
Unfortunatelly it still does get collected.

At least now the user gets a warning when the old add method is used and he is asked to switch to the createFunction method which seems to be safe.