Two of the problems with the R bindings have been solved using typemaps. I actually think these are bugs in SWIG.
The last problem that I could not solve yet is the problem that wrapped classes
are not cast to the most specific type.
Everything I tried so far has not lead to any success. I probably don't know enough about the
R internals and I have not found an example of how to do this yet.