Kicked out all the typecasting coe for the R bindings because R normally does not seem to need it. I wrote a small example that was very well able to cast objects to the most specific type automatically. This would also explain the lack of documentation and examples.
I still don't know for sure why this does not seem to work for the COPASI classes.
Right now my guess would be that the multiple inheritance we are using throws SWIG off.

I have now addded a hack to CReaction.i to be able to use the function objects returned
by CFunctionDB::findFunction in reactions.