Change handling of outputStartTime for scans
In the case of useInitialValues=false, the output is not started after
modeltime==outputStartTime, but after a delay of "outputStartTime" with
respect to the begin of the simulation.
Reason: When using a scan with "always apply initial conditions" set to
false, it should be possible to delay output for each iteration, not
just for the first.