- issue 2967: expose converters to swig, allow to customize output generated
also added documentation to the exporter, and cleaned up a number of statements that had no influence on the export.

added an overload to CPlotSpecification::appliesTo, so that it is easy to ask whether it a applies to a given task (rather than having to construct a container list first).